fix: move handler code into index
There was an issue passing certain request/response variables from the main script, and with the new import scheme there isn't as much code bloat at the top.
